edit_common: add file for edit* functions
These functions were being used in different places(commands), thus having them
declared in other place (a common file) other than in a specific issue_ command
seems better.
